+++ tests/core/test_storage.py
@@ -342,6 +342,27 @@ class BaseTestStorage(object):
             **self.storage_kw
         )
 
+    def test_get_all_handles_parent_id_pattern_matching(self):
+        self.create_record(parent_id='abc', collection_id='c')
+        record = self.create_record(parent_id='abc', collection_id='c')
+        self.storage.delete(object_id=record['id'], parent_id='abc', collection_id='c')
+        self.create_record(parent_id='efg', collection_id='c')
+
+        records, total_records = self.storage.get_all(parent_id='ab*', collection_id='c',
+                                                      include_deleted=True)
+        self.assertEquals(len(records), 2)
+        self.assertEquals(total_records, 1)
+
+    def test_get_all_does_proper_parent_id_pattern_matching(self):
+        self.create_record(parent_id='abc', collection_id='c')
+        self.create_record(parent_id='xabcx', collection_id='c')
+        self.create_record(parent_id='efg', collection_id='c')
+
+        records, total_records = self.storage.get_all(parent_id='ab*', collection_id='c',
+                                                      include_deleted=True)
+        self.assertEquals(len(records), 1)
+        self.assertEquals(len(records), total_records)
+
     def test_get_all_return_all_values(self):
         for x in range(10):
             record = dict(self.record)
@@ -853,6 +874,24 @@ class DeletedRecordsTest(object):
                                               include_deleted=True)
         self.assertEqual(count, 0)
         self.assertEqual(len(records), 0)
+        records, count = self.storage.get_all(parent_id='efg',
+                                              collection_id='c',
+                                              include_deleted=True)
+        self.assertEqual(count, 1)
+        self.assertEqual(len(records), 1)
+
+    def test_delete_all_does_proper_parent_id_matching(self):
+        self.create_record(parent_id='abc', collection_id='c')
+        self.create_record(parent_id='xabcx', collection_id='c')
+        self.create_record(parent_id='efg', collection_id='c')
+        self.storage.delete_all(parent_id='ab*',
+                                collection_id=None,
+                                with_deleted=False)
+        records, count = self.storage.get_all(parent_id='xabcx',
+                                              collection_id='c',
+                                              include_deleted=True)
+        self.assertEqual(count, 1)
+        self.assertEqual(len(records), 1)
         records, count = self.storage.get_all(parent_id='efg',
                                               collection_id='c',
                                               include_deleted=True)